lectures/requirements.txt
ipykernel
matplotlib
mujoco
mujoco-mjx
jax
judo-rai
plotly